operations/CDK/OperationsPrerequisites/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.230
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==41.2.0
    # via aws-cdk-lib
aws-cdk-lib==2.188.0
    # via -r requirements.in
boto3==1.37.27
    # via -r requirements.in
botocore==1.37.27
    # via
    #   boto3
    #   s3transfer
cattrs==24.1.3
    # via jsii
constructs==10.4.2
    # via
    #   -r requirements.in
    #   aws-cdk-lib
importlib-resources==6.5.2
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.111.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
pytz==2025.2
    # via -r requirements.in
s3transfer==0.11.4
    # via boto3
six==1.17.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.13.1
    # via jsii
urllib3==2.3.0
    # via botocore
